apps/wiki-server/requirements.txt
flask==3.0.3
pyyaml==6.0.2
markdown==3.7
pygments==2.18.0

